Fixed Button Hover UI
On entering a button, a sound will play and the button will be highlighted in a grey color. When clicking the button, a different sound will play and the button's color will turn darker. When exiting a button, the button will go back to it's regular color. You can also use the arrow keys and enter to navigate the menu, except inside of the settings menu.